requirements.txt
accelerate
datasets
numpy
pandas
transformers
torchdata
deepspeed
wandb
hydra-core